Added the supported metering variable to PlayLayer
From my findings, legacy levels are indicated if this bool is false. If it is, GD will default to extracting the metering from the AudioEffectsLayer rather than using GameSoundManager.

I say this is likely a way of indicating a legacy mode because this can be found in all levels which were made before 2.1.